Tidy up Encoder, IMU, Send and Receive modules

+++ Firmware/Encoder/Encoder.h
@@ -1,17 +1,24 @@
 /*
- * File:   Encoder.h
- * Author: Seb Madgwick
+ * @file Encoder.h
+ * @author Seb Madgwick
+ * @brief Quadrature encoder module to provide measurements of rotational
+ * position and velocity.
  */
 
 #ifndef ENCODER_H
 #define ENCODER_H
 
 //------------------------------------------------------------------------------
+// Includes
+
+#include <stdint.h> // uint16_t, int16_t, int32_t
+
+//------------------------------------------------------------------------------
 // Function prototypes
 
 void EncoderInitialise();
-long EncoderGetPostion();
-long EncoderGetVelocity();
+int32_t EncoderGetPostion();
+int32_t EncoderGetVelocity();
 void EncoderZero();
 
 #endif
